backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
PyJWT
python-multipart
google-generativeai
pydantic
python-dotenv
psycopg2-binary